Makefile ahora demora menos... resulta que llamaba automaticamente a todas las clases que necesitaba compilar... o sea, compilabamos mas de una vez cada una. Se cambio el metodo de inicializacion de los paneles, ahora se instancian y luego se llama al metodo init(). Por mientras el screen panel y el tracking permiten ser controlados por el mismo software y por ACS... hay que ver cuan factible es en realidad que funcionara asi.
git-svn-id: https://csrg.inf.utfsm.cl/repos/software/CSAT@262 75a0fe40-ac5b-440f-90d5-6af9f3fbf44f

+++ Hevelius/src/trunk/Hevelius/virtualview/VirtualTelescope.java
@@ -9,7 +9,7 @@ import java.awt.*;
 
 public class VirtualTelescope
 {
-	public static Image getScreen(String RA, String DEC, float x, float y)
+	public static Image getScreen(String RA, String DEC, double x, double y)
 	{
 		try
 		{
